split files & some backlog fixes
