+++ main/helpers/scraping/hashes.ts
