+++ main/helpers/scraping/response-transform.ts
